+++ app/build.gradle
@@ -13,9 +13,13 @@ android {
         vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
+        debug {
+            resValue "string", "google_maps_key", (project.findProperty("GOOGLE_MAPS_API_KEY") ?: "")
+        }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            resValue "string", "google_maps_key", (project.findProperty("GOOGLE_MAPS_API_KEY") ?: "")
         }
     }
     compileOptions {
@@ -27,8 +31,6 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:support-v4:26.1.0'
-    implementation 'com.google.android.gms:play-services-maps:12.0.0'
-    compile "com.google.android.gms:play-services-location:12.0.0"
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
     implementation 'com.jakewharton:butterknife:8.8.1'
     implementation 'com.android.support:design:26.1.0'
@@ -41,6 +43,11 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
     implementation 'com.android.support:cardview-v7:26.1.0'
     implementation 'com.android.support:support-vector-drawable:26.1.0'
+
+    // Location
+    implementation "com.google.android.gms:play-services-places:12.0.1"
+    implementation 'com.google.android.gms:play-services-maps:12.0.1'
+    compile "com.google.android.gms:play-services-location:12.0.1"
 
     //Butterknife
     implementation 'com.jakewharton:butterknife:8.8.1'
