--- app/src/debug/res/values/google_maps_api.xml
@@ -1,24 +1,0 @@
-<resources>
-    <!--
-    TODO: Before you run your application, you need a Google Maps API key.
-
-    To get one, follow this link, follow the directions and press "Create" at the end:
-
-    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=32:0B:0D:8E:2D:EE:E1:BB:4F:2B:3D:0E:DA:FF:1D:CA:43:97:03:CE%3Bes.udc.apm.familycare
-
-    You can also add your credentials to an existing key, using these values:
-
-    Package name:
-    32:0B:0D:8E:2D:EE:E1:BB:4F:2B:3D:0E:DA:FF:1D:CA:43:97:03:CE
-
-    SHA-1 certificate fingerprint:
-    32:0B:0D:8E:2D:EE:E1:BB:4F:2B:3D:0E:DA:FF:1D:CA:43:97:03:CE
-
-    Alternatively, follow the directions here:
-    https://developers.google.com/maps/documentation/android/start#get-key
-
-    Once you have your key (it starts with "AIza"), replace the "google_maps_key"
-    string in this file.
-    -->
-    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">YOUR_KEY_HERE</string>
-</resources>
