+++ app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
          location permissions for the 'MyLocation' functionality. 
     -->
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+    <uses-permission android:name="android.permission.INTERNET" />
 
 
     <application
@@ -52,7 +53,7 @@
         -->
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="AIzaSyCSxEFzvOYx6X_ihn6fXQA1yaKjpYA0AaE" />
+            android:value="@string/google_maps_key" />
 
     </application>
