+++ gradle.properties
@@ -15,4 +15,4 @@ org.gradle.jvmargs=-Xmx1536m
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-GOOGLE_MAPS_API_KEY=AIzaSyCSxEFzvOYx6X_ihn6fXQA1yaKjpYA0AaE
+GOOGLE_MAPS_API_KEY=AIzaSyBA4r7jQxoCUJR1sxmH939zjgQY7E4bxdo
